refactor(nika-builtin): net_payload split — the 1500-LOC file cap holds
The form/multipart assembly (prepare_request · form_body · the closed
part resolver) + its tests move to net_payload.rs: net.rs was at 1620
post-vNext (cap 1500 · vector 12). Transport + extraction stay in
net.rs · the crawl stays in net_traverse.rs — one owner per concern.
Also fixes the digest.rs intra-doc link (crate::ExtractMode · vector
28). Zero behavior change: 276+136 lib tests · clippy 0 · doc 0.
